restore: handle unsuffixed TiDB version (fix TOOL-360) (#67)
TiDB versions are generated with `git describe --tags`. If the release
is made exactly on tag `vX.Y.Z-P`, the version will be
`5.7.10-TiDB-vX.Y.Z-P`. If the release is not a commit exactly
corresponding to a tag, the description will be like
`vX.Y.Z-P-99-g123abcd`. This patch assumes TiDB never makes a release
with multiple patch versions (`vX.Y.Z-P1-P2`).