Update dependencies and remove juju/errors (#79)
* *: replaced juju/errors by pingcap/errors and use builtin context package

* vendor: update tidb to v2.1.0-rc.3 and update its dependencies

* vendor: update the rest of the direct dependencies

* license: added commercial license

* vendor: update TiDB and PD to v2.1.0-rc.4

Removed the goyacc step since the parser has been moved out. For now we
rely on `git checkout` to recover the parser until v3.0.0-alpha is tagged

* tests: reduce the number of max-open-files used by tikv-server

Jenkin's open file limit is somehow reduced to 65536, and the default tikv
requirement is 40960*2+1000 = 82920, causing CI failing to start the
integration tests. Here we reduce the number to 4096*2+1000 = 9192 via
explicit config.

* license: fix the year range